+++ ChangeLog.md
@@ -18,9 +18,16 @@ to browse the changes between the tags.
 
 See docs/process.md for more on how version tagging works.
 
-4.0.23 (in development)
+4.0.24 (in development)
 -----------------------
 - libunwind was updated to LLVM 21.1.8. (#26035)
+- compiler-rt was updated to LLVM 21.1.8. (#26405)
+- A new `-sEXECUTABLE` setting was added which adds a #! line to the resulting
+  JavaScript and makes it executable.  This setting defaults to true when the
+  output filename has no extension, or ends in `.out` (e.g. `a.out`) (#26085)
+
+4.0.23 - 01/10/26
+-----------------
 - The inconsistency of incrementing / decrementing refcounts between Wasm EH and
   Emscripten EH has been fixed. See `test_EXPORT_EXCEPTION_HANDLING_HELPERS` in
   `test_core.py` to see the usage. (#25988)
@@ -34,13 +41,14 @@ See docs/process.md for more on how version tagging works.
   have a fetch API, you will need to either use `-sSINGLE_FILE` (to embed the
   Wasm file), or use a custom `instantiateWasm` callback to supply the
   Wasm module yourself. (#25942)
+- Upgraded `contrib.lua` to Lua version 5.5.0 (#26033)
 
 4.0.22 - 12/18/25
 -----------------
 - Source maps now support 'names' field with function name information.
   emsymbolizer will show function names when used with a source map. The size
   of source maps may increase 2-3x and the link time can increase slightly due
-  to more processing on source map creation. (#25298)
+  to more processing on source map creation. (#25928)
 - Emscripten will now cache the JS code that it generates and re-use when
   linking with the same settings at a later date.  This should improve link
   times generally but should especially noticeable when linking lots of small
@@ -237,7 +245,7 @@ See docs/process.md for more on how version tagging works.
 - When JSPI is enabled `async` library functions are no longer automatically
   wrapped with `WebAssembly.Suspending` functions. To automatically wrap library
   functions for use with JSPI they must now explicitly set
-  `myLibraryFunction__async: true`.
+  `myLibraryFunction__async: true`. (#24550)
 - Removed special casing for `size_t` in Embind, since it was also inadvertently
   affecting `unsigned long` on wasm64. Both will now match the behaviour of
   other 64-bit integers on wasm64 and will be passed as `bigint` instead of
@@ -656,7 +664,7 @@ See docs/process.md for more on how version tagging works.
 - Fix the location of the dummy `.worker.js` file that is now generated as part
   of pthread builds so that is generated alongside the main JavaScript file.
   See #21701. ()
-- `-sASYNCIFY=2` is setting now deprecated, use `-sJSPI` instead.
+- `-sASYNCIFY=2` is setting now deprecated, use `-sJSPI` instead. (#21824)
 
 3.1.58 - 04/23/24
 -----------------